pyproject.toml
[project]
name = "pixelcare"
version = "0.1.0"
description = "AI health companion via webcam"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "opencv-python>=4.8.0",
    "mediapipe>=0.10.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "gradio>=4.0.0",
    "openai>=1.0.0",
    "toml>=0.10.2",
    "pypdf2>=3.0.1",
    "pillow>=10.4.0",
    "pdf2image>=1.17.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
